[ews-build.webkit.org] Switch macOS JSC arm64 stress test queue to O3 debug builds rdar://159975164 https://bugs.webkit.org/show_bug.cgi?id=303943
Reviewed by Keith Miller.

This adds an assert-enabled JSC stress test configuration to pre-commit testing so that regressions
can be caught before landing. This configuration has been running successfully in post-commit
for the past few weeks.

Switching to O3 debug builds adds ~20 minutes to test runtime on our existing arm64 EWS hardware
(51m vs. 31m on release builds). The x86 configuration will not change at this time.

* Tools/CISupport/Shared/steps.py:
(SetO3OptimizationLevel): Move this step to share it between ews-build and build.webkit.og
* Tools/CISupport/build-webkit-org/steps.py:
(SetO3OptimizationLevel): Deleted.
* Tools/CISupport/ews-app/ews/common/github.py:
(GitHubEWS): Update queue name to reflect the configuration change.
* Tools/CISupport/ews-app/ews/views/statusbubble.py:
(StatusBubble): Ditto.
* Tools/CISupport/ews-build/config.json: Ditto.
* Tools/CISupport/ews-build/factories.py:
(JSCBuildO3AndTestsFactory): Add a new factory that includes the step to set the O3 optimization level.
* Tools/CISupport/ews-build/factories_unittest.py:
(TestExpectedBuildSteps): Update unit tests.
* Tools/CISupport/ews-build/loadConfig.py: Import the SetO3OptimizationLevel step.
* Tools/CISupport/ews-build/steps.py: Ditto.
(RunJavaScriptCoreTests.evaluateCommand): Run the SetO3OptimizationLevel step before re-building
without the PR.

Canonical link: https://commits.webkit.org/305715@main
